stock-analysis/requirements.airflow.txt
# requirements.airflow.txt
apache-airflow>=2.7.1
pandas
yfinance
pyarrow
aiohttp
asyncio